move cost stuff to deterministic
